templates working without if/list, add data-value support
This changes frees the template compiler of the runtime polyfills for `instantiate="if"` and `iterate`. With the new emitter framework, it was easy to add codegen for iterate.

The one caveat is I have temporarily placed most of the app inside a webcomponent. I expect this will "just work" once the "main" HTML file is built on the new Emitter architecture.

misc changes:
* delete if/list from js_polyfill.
* reverted an old workaround in web_components for if/list.
* change to data-value syntax and implement support for it in the analyzer.
* fix computing the set of used web_components: it was only finding the ones used in main. This computation now happens in the analyzer.
* tests for the new analyzer features.
* fix a bug where "if" was not calling created in the right order.
* expand the web component declarations before injecting into the document, so .created can be called at creation time instead of when the node is inserted.

Review URL: https://chromiumcodereview.appspot.com//10905252