Noticed our tests were getting slow. By keeping all of them in one Dart process, it cuts our runtime almost in half: from 6.6s -> 3.7s on my machine.
The remaining time is mostly in the shell script compare step. My suspicion is DumpRenderTree. We probably need to parallelize it, or do a similar trick of running multiple tests in one DRT process.

It is one extra file to maintain (and hardcoded list of _test.dart files), but IMO it is worth it for almost double edit+test cycles.

Review URL: https://chromiumcodereview.appspot.com//10991050