This change makes it possible to run "dwc" from another package that imports web_components. In particular, all interesting logic moves into a public library instead of being in the "bin/dwc" file.
I also fixed build.dart to work for .html files. However, I disabled it by default as this generates a lot of clutter in the tree. We need a reliable "--clean" before we can turn it on, IMO.

Also changed the convention to generate file names with the underscore prefix. This groups generated code nicely in the Editor.

Review URL: https://chromiumcodereview.appspot.com//11077005