fixes #203 - return proper exit code
We need to use transformException instead of handleException. The latter breaks the future chain, causing the FutureGroup to never complete. But the VM exits because it has no more pending operations. Doh! Thanks to Nathan for quickly spotting this (he noted that "handleException" is a code smell and should only be used by forwarding methods).

I also noticed that FutureGroup should "return true" in its handleException since it is forwarding the exception to another completer.

Review URL: https://chromiumcodereview.appspot.com//11421090